Fix slides lint
