+++ app/components/home/slides/HomeSlide2.js
@@ -12,7 +12,7 @@ function HomeSlide2() {
           link: '/global-scenarios/32.47/-50.10/3?maps=0,climate,avg-precipitation'
         }
       ]}
-      />
+    />
   );
 }
